enable GPIO0 33 and 36 in device tree
GPIO0 shows up as gpiochip1. To correctly understand the chip number,
do `gpiodetect' command and compare the address printed with the base
dtsi file. With the right chip number determined, one can toggle the
GPIO pins as follows:

gpioset gpiochip1 33=1
gpioset gpiochip1 33=0

(replace 33 with the GPIO pin number and the gpiochip1 with the
correct chip determined from schematics and the above method)